refator: extracted table view of import data to component
